# party-amea/party-model/src/main/resources/frontend/swagger_party_obw.json
{
   "swagger":"2.0",
   "info":{
      "description":"ajouter des attributs mandatory pour renseigner le payment Method.\ncomme solution provisoire on va gérer ces méthodes de payment dans cette API. puisque OBW n'expose pas l'API PaymentMethod.",
      "version":"0.1.2_tag",
      "title":"API PartymanagementNotification",
      "x-logo":{
         "url":"\/redoc\/logo.png",
         "backgroundColor":"#FFFFFF"
      }
   },
   "produces":[
      "application\/json;charset=utf-8"
   ],
...
   "paths":{
...
      "\/individualCreateEvent":{
         "post":{
            "tags":[
               "IndividualCreateEvent"
            ],
            "consumes":[
               "application\/json;charset=utf-8"
            ],
            "produces":[
               "application\/json;charset=utf-8"
            ],
            "operationId":"individualCreateEventCreate",
            "summary":"create individualCreateEvent",
            "deprecated":false,
            "parameters":[
               {
                  "name":"IndividualCreateEvent",
                  "required":true,
                  "in":"body",
                  "schema":{
                     "$ref":"#\/definitions\/IndividualCreateEvent"
                  }
               }
            ],
            "responses":{
               "201":{
                  "description":"Created",
                  "schema":{
                     "$ref":"#\/definitions\/IndividualCreateEvent"
                  }
               },
               "400":{
                  "description":"Bad Request\n\nList of supported error codes:\n- 20: Invalid URL parameter value\n- 21: Missing body\n- 22: Invalid body\n- 23: Missing body field\n- 24: Invalid body field\n- 25: Missing header\n- 26: Invalid header value\n- 27: Missing query-string parameter\n- 28: Invalid query-string parameter value",
                  "schema":{
                     "$ref":"#\/definitions\/Error"
                  }
               },
               "404":{
                  "description":"Not Found\n\nList of supported error codes:\n- 60: Resource not found",
                  "schema":{
                     "$ref":"#\/definitions\/Error"
                  }
               },
               "422":{
                  "description":"Unprocessable entity\n\nFunctional error",
                  "schema":{
                     "$ref":"#\/definitions\/Error"
                  }
               },
               "503":{
                  "description":"Service Unavailable\n\nList of supported error codes:\n- 5: The service is temporarily unavailable\n- 6: Orange API is over capacity, retry later !",
                  "schema":{
                     "$ref":"#\/definitions\/Error"
                  }
               }
            }
         }
      },
      "\/individualAttributeValueChangeEvent":{
         "post":{
            "tags":[
               "IndividualAttributeValueChangeEvent"
            ],
            "consumes":[
               "application\/json;charset=utf-8"
            ],
            "produces":[
               "application\/json;charset=utf-8"
            ],
            "operationId":"individualAttributeValueChangeEventCreate",
            "summary":"create individualAttributeValueChangeEvent",
            "deprecated":false,
            "parameters":[
               {
                  "name":"IndividualAttributeValueChangeEvent",
                  "required":true,
                  "in":"body",
                  "schema":{
                     "$ref":"#\/definitions\/IndividualAttributeValueChangeEvent"
                  }
               }
            ],
            "responses":{
               "201":{
                  "description":"Created",
                  "schema":{
                     "$ref":"#\/definitions\/IndividualAttributeValueChangeEvent"
                  }
               },
               "400":{
                  "description":"Bad Request\n\nList of supported error codes:\n- 20: Invalid URL parameter value\n- 21: Missing body\n- 22: Invalid body\n- 23: Missing body field\n- 24: Invalid body field\n- 25: Missing header\n- 26: Invalid header value\n- 27: Missing query-string parameter\n- 28: Invalid query-string parameter value",
                  "schema":{
                     "$ref":"#\/definitions\/Error"
                  }
               },
               "404":{
                  "description":"Not Found\n\nList of supported error codes:\n- 60: Resource not found",
                  "schema":{
                     "$ref":"#\/definitions\/Error"
                  }
               },
               "422":{
                  "description":"Unprocessable entity\n\nFunctional error",
                  "schema":{
                     "$ref":"#\/definitions\/Error"
                  }
               },
               "503":{
                  "description":"Service Unavailable\n\nList of supported error codes:\n- 5: The service is temporarily unavailable\n- 6: Orange API is over capacity, retry later !",
                  "schema":{
                     "$ref":"#\/definitions\/Error"
                  }
               }
            }
         }
      },
      "\/organizationCreateEvent":{
         "post":{
            "tags":[
               "OrganizationCreateEvent"
            ],
            "consumes":[
               "application\/json;charset=utf-8"
            ],
            "produces":[
               "application\/json;charset=utf-8"
            ],
            "operationId":"organizationCreateEventCreate",
            "summary":"create organizationCreateEvent",
            "deprecated":false,
            "parameters":[
               {
                  "name":"OrganizationCreateEvent",
                  "required":true,
                  "in":"body",
                  "schema":{
                     "$ref":"#\/definitions\/OrganizationCreateEvent"
                  }
               }
            ],
            "responses":{
               "201":{
                  "description":"Created",
                  "schema":{
                     "$ref":"#\/definitions\/OrganizationCreateEvent"
                  }
               },
               "400":{
                  "description":"Bad Request\n\nList of supported error codes:\n- 20: Invalid URL parameter value\n- 21: Missing body\n- 22: Invalid body\n- 23: Missing body field\n- 24: Invalid body field\n- 25: Missing header\n- 26: Invalid header value\n- 27: Missing query-string parameter\n- 28: Invalid query-string parameter value",
                  "schema":{
                     "$ref":"#\/definitions\/Error"
                  }
               },
               "404":{
                  "description":"Not Found\n\nList of supported error codes:\n- 60: Resource not found",
                  "schema":{
                     "$ref":"#\/definitions\/Error"
                  }
               },
               "422":{
                  "description":"Unprocessable entity\n\nFunctional error",
                  "schema":{
                     "$ref":"#\/definitions\/Error"
                  }
               },
               "503":{
                  "description":"Service Unavailable\n\nList of supported error codes:\n- 5: The service is temporarily unavailable\n- 6: Orange API is over capacity, retry later !",
                  "schema":{
                     "$ref":"#\/definitions\/Error"
                  }
               }
            }
         }
      },
      "\/organizationAttributeValueChangeEvent":{
         "post":{
            "tags":[
               "OrganizationAttributeValueChangeEvent"
            ],
            "consumes":[
               "application\/json;charset=utf-8"
            ],
            "produces":[
               "application\/json;charset=utf-8"
            ],
            "operationId":"organizationAttributeValueChangeEventCreate",
            "summary":"create organizationAttributeValueChangeEvent",
            "deprecated":false,
            "parameters":[
               {
                  "name":"OrganizationAttributeValueChangeEvent",
                  "required":true,
                  "in":"body",
                  "schema":{
                     "$ref":"#\/definitions\/OrganizationAttributeValueChangeEvent"
                  }
               }
            ],
            "responses":{
               "201":{
                  "description":"Created",
                  "schema":{
                     "$ref":"#\/definitions\/OrganizationAttributeValueChangeEvent"
                  }
               },
               "400":{
                  "description":"Bad Request\n\nList of supported error codes:\n- 20: Invalid URL parameter value\n- 21: Missing body\n- 22: Invalid body\n- 23: Missing body field\n- 24: Invalid body field\n- 25: Missing header\n- 26: Invalid header value\n- 27: Missing query-string parameter\n- 28: Invalid query-string parameter value",
                  "schema":{
                     "$ref":"#\/definitions\/Error"
                  }
               },
               "404":{
                  "description":"Not Found\n\nList of supported error codes:\n- 60: Resource not found",
                  "schema":{
                     "$ref":"#\/definitions\/Error"
                  }
               },
               "422":{
                  "description":"Unprocessable entity\n\nFunctional error",
                  "schema":{
                     "$ref":"#\/definitions\/Error"
                  }
               },
               "503":{
                  "description":"Service Unavailable\n\nList of supported error codes:\n- 5: The service is temporarily unavailable\n- 6: Orange API is over capacity, retry later !",
                  "schema":{
                     "$ref":"#\/definitions\/Error"
                  }
               }
            }
         }
      }
   },
...
}